+++ src/data_preparation/data_preparation.py
@@ -2,8 +2,8 @@ import os
 from typing import Any, Dict, List, Tuple
 
 from constants_ import ACQUISITION_COLS, PERFORMANCE_COLS
-from src.imputation.central_imputation import CentralImputation
-from src.utils import get_dataframe_summary, is_nvidia_gpu_available, timeit
+from imputation.central_imputation import CentralImputation
+from utils import get_dataframe_summary, is_nvidia_gpu_available, timeit
 
 NVIDIA_GPU_AVAILABILITY = is_nvidia_gpu_available()
 
@@ -158,4 +158,3 @@ class DataPreparation:
         print("Shape:", df.shape)
         df = CentralImputation(NVIDIA_GPU_AVAILABILITY).impute(df)
         return df
-
